requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.4
attrs==25.3.0
audioop-lts==0.2.2       
brotli==1.2.0
browser-cookie3==0.20.1  
certifi==2025.10.5       
charset-normalizer==3.4.4
discord.py==2.6.4        
ffmpeg-python==0.2.0     
frozenlist==1.7.0        
future==1.0.0
greenlet==3.2.4
idna==3.10
lz4==4.4.5
multidict==6.6.4
mutagen==1.47.0
packaging==24.2
pefile==2023.2.7
playwright==1.56.0
propcache==0.3.2
pycryptodomex==3.23.0
pyee==13.0.0
pygame==2.6.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
pywin32==311
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==78.1.0
shadowcopy==0.0.4
shiboken6==6.10.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
WMI==1.5.1
yarl==1.20.1
yt-dlp==2025.10.22